code/my_ocr_pipeline/requirements.txt
transformers==4.28.1
PyPDF2==3.0.1
python-docx==0.8.11
pytesseract==0.3.10
Pillow==9.5.0
pdf2image==1.16.3
fastapi==0.95.2
uvicorn==0.22.0
